Fix potential hang when automatically logging in because a service task needs credentials.
Previously, could enter a state where a service task was called, automatically redirected to authenticate, but was then discarded before it could execute the task.

Now, if we log out we also explicitly de-authenticate geocode and route service tasks, and we don’t replace geocode and route service tasks on login if they’re the same URL as the new (logged-in) portal.